% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychometry.R
\name{difficultyindex}
\alias{difficultyindex}
\alias{difficulty_index}
\title{Difficulty index}
\usage{
difficultyindex(x, success)
}
\arguments{
\item{x}{a vector that records if the answers are correct or wrong.}

\item{success}{the (only) value taken as correct answer.}
}
\value{
The computed difficulty index.
}
\description{
Difficulty index definition for binary variables and two level factors.
}
\details{
Computes the difficulty index for binary variables or two level factor.
It can also be applied to any numeric or factor variable taking into account that any value that differs from the given as success will be considered as failure.
This includes NA values.
}
