% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsySummaryMenu.R, R/mvnMenu.R
\name{numerical_summaries}
\alias{numerical_summaries}
\alias{PsySummaryMenu}
\alias{Numerical_summaries}
\alias{Numerical_item_analysis}
\alias{mvnMenu}
\title{Numerical summaries}
\usage{
PsySummaryMenu()

mvnMenu()
}
\value{
The abovementioned statistics.

The abovementioned statistics.
}
\description{
Calculate some numerical summaries for items

Calculate some numerical summaries for items
}
\details{
From a set of question scores (at least two) and a test scores, this function calls to \code{\link[psychometric]{item.exam}} from psychometric packages.

Its description in psychometric package documentation is:
Conducts an item level analysis. Provides item-total correlations, Standard deviation in items, difficulty, discrimination, and reliability and validity indices.

To use this function from R-Commander, select from menu "Psychometry" -> "Item analysis..." -> "Numerical summaries..."

From a set of question scores (at least two) and a test scores, this function calls to \code{\link[psychometric]{item.exam}} from psychometric packages.

Its description in psychometric package documentation is:
Conducts an item level analysis. Provides item-total correlations, Standard deviation in items, difficulty, discrimination, and reliability and validity indices.

To use this function from R-Commander, select from menu "Psychometry" -> "Item analysis..." -> "Numerical summaries..."
}
\seealso{
\code{\link[psychometric]{item.exam}}

\code{\link[psychometric]{item.exam}}
}
