\name{qccMenu.es}
\title{Opcioned de menú para hacer gráficos de control de calidad}
\alias{Gráficos de control de calidad (es)}
\docType{package}
\encoding{UTF-8}
\description{
  Dentro del menú "Control de Calidad" se proporcionan nuevas entradas
  para realizar múltipes gráficos de control de calidad.
}

\details{
  Estas opciones usan la función \code{qcc} del paquete \code{qcc}.
  Para más información véase \code{\link[qcc]{qcc}}.
}

\author{
  Manuel Munoz-Marquez <manuel.munoz@uca.es>
}

\keyword{package}

\seealso{
  Para más información véase \code{\link[qcc]{qcc}} y \code{\link[Rcmdr]{Rcmdr-package}}. 
}
