\name{qccMenu}
\title{Menu options to make a quality control graphs}
\alias{qccMenu}
\alias{Quality control graphs}

\alias{c-chart}
\alias{np-chart}
\alias{p-chart}
\alias{R-chart}
\alias{S-chart}
\alias{u-chart}
\alias{xbar-chart}

\docType{package}
\encoding{UTF-8}
\description{
  Within the menu "Quality Control" new entries are provided to make mutiple
  quality control graphs.}

\details{
  These options use the \code{qcc} function from the \code{qcc} package.
  For more information see \code{\link[qcc]{qcc}}.
}

\author{
  Manuel Munoz-Marquez <manuel.munoz@uca.es>
}

\keyword{package}

\seealso{
  For more information see \code{\link[qcc]{qcc}} and \code{\link[Rcmdr]{Rcmdr-package}}. 

  Para ayuda en \enc{español}{espanol}, \enc{véase}{vease}
  \code{\link{Gráficos de control de calidad (es)}}. (For Spanish help see
  \code{\link{Gráficos de control de calidad (es)}}.) 
}
