\name{RcmdrPlugin.UCA-internal}
\title{RcmdrPlugin.UCA: Some extension to Rcmdr (R Commander) made by
  R-UCA project and used in teaching statistics at University of Cadiz
  (UCA).}

\alias{Binaries}
\alias{BinariesP}
\alias{cchart1Menu}
\alias{cchart2Menu}
\alias{cchartpMenu}
\alias{Dicotomics}
\alias{DicotomicsP}
\alias{input2predict}
\alias{NIAMenu}
\alias{npchart1Menu}
\alias{npchart2Menu}
\alias{npchartpMenu}
\alias{numeric.runs.test}
\alias{pchart1Menu}
\alias{pchart2Menu}
\alias{pchartpMenu}
\alias{predict4dataset}
\alias{PsyAlphaCIMenu}
\alias{PsyAlphaMenu}
\alias{PsyCDIMenu}
\alias{PsyDIMenu}
\alias{PsyGuttmanMenu}
\alias{PsySaturationMenu}
\alias{PsySBMenu}
\alias{PsySBReliabilityMenu}
\alias{PsySBSizeMenu}
\alias{PsyScoreMenu}
\alias{R1mcMenu}
\alias{R2mcMenu}
\alias{RpmcMenu}
\alias{randomnessFTest}
\alias{randomnessNTest}
\alias{S1mcMenu}
\alias{S2mcMenu}
\alias{SpmcMenu}
\alias{SampleSizeZMenu}
\alias{SDIMenu}
\alias{sigma.test}
\alias{sigmaTest}
\alias{twolevelfactor.runs.test}
\alias{uchart1Menu}
\alias{uchart2Menu}
\alias{uchartpMenu}
\alias{xbar1mcMenu}
\alias{xbar2mcMenu}
\alias{xbarpmcMenu}
\alias{xbarone1Menu}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  RcmdrPlugin.UCA packages. 
  For user information ask for help about \code{\link[RcmdrPlugin.UCA]{RcmdrPlugin.UCA}}.}
\keyword{internal}
