% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsySummaryMenu.R, R/mvnMenu.R
\name{Resúmenes_numéricos}
\alias{Resúmenes_numéricos}
\title{Resúmenes numéricos}
\value{
Los estadísticos mencionados anteriormente.

Los estadísticos mencionados anteriormente.
}
\description{
Calcular algunos resúmenes numéricos para los ítems.

Calcular algunos resúmenes numéricos para los ítems.
}
\details{
Desde un conjunto de puntaciones de preguntas (al menos dos) y de test, esta función llama a \code{\link[psychometric]{item.exam}} del paquete phychometric.

Su descripción, en inglés, en la documentación del paquete psychometric es:
Realiza un análisis a nivel de elemento. Proporciona correlaciones ítem-total, Desviación estándar en ítems, dificultad, discriminación e índices de confiabilidad y validez.

Para usar esta función desde R-Commander, seleccione del menú "Psicometría" -> "Análisis de ítems..." -> "Resúmenes numéricos..."

Desde un conjunto de puntaciones de preguntas (al menos dos) y de test, esta función llama a \code{\link[psychometric]{item.exam}} del paquete phychometric.

Su descripción, en inglés, en la documentación del paquete psychometric es:
Realiza un análisis a nivel de elemento. Proporciona correlaciones ítem-total, Desviación estándar en ítems, dificultad, discriminación e índices de confiabilidad y validez.

Para usar esta función desde R-Commander, seleccione del menú "Psicometría" -> "Análisis de ítems..." -> "Resúmenes numéricos..."
}
\seealso{
\code{\link[psychometric]{item.exam}}

\code{\link[psychometric]{item.exam}}
}
