% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychometry.R
\name{indicedificultad}
\alias{indicedificultad}
\alias{Índice_de_dificultad}
\title{Índice de dificultad}
\arguments{
\item{x:}{un vector que indica si las respuestas son correctas o incorrectas.}

\item{success:}{el (único) valor corisderado como correcto.}
}
\value{
El índice de dificultad calculado.
}
\description{
Definición del índice de dificultad para variables binarias y factores de dos niveles.
}
\details{
Calcula el índice de dificultad para variables binarias o factor de dos niveles.
También se puede aplicar a cualquier variable numérica o factorial teniendo en cuenta que cualquier valor que difiera del dado como éxito será considerado como fracaso.
Esto incluye valores NA.
}
