\name{RcmdrPlugin.UCA-internal}
\title{RcmdrPlugin.UCA: Some extension to Rcmdr (R Commander) made by R-UCA project and used in teaching statistics at University of Cadiz (UCA).}
\alias{numeric.runs.test}
\alias{input2predict}
\alias{predict4dataset}
\alias{randomnessFTest}
\alias{randomnessNTest}
\alias{twolevelfactor.runs.test}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  RcmdrPlugin.UCA packages. 
  For user information ask for help about \code{\link[RcmdrPlugin.UCA]{RcmdrPlugin.UCA}}.}
\keyword{internal}
