\name{RcmdrPlugin.UCA-internal}
\title{RcmdrPlugin.UCA: Some extension to Rcmdr (R Commander) made by
  R-UCA project and used in teaching statistics at University of Cadiz
  (UCA).}

\alias{cchart12Menu}
\alias{cchart1Menu}
\alias{cchart2Menu}
\alias{input2predict}
\alias{npchart1Menu}
\alias{npchart12Menu}
\alias{npchart2Menu}
\alias{numeric.runs.test}
\alias{pchart1Menu}
\alias{pchart12Menu}
\alias{pchart2Menu}
\alias{predict4dataset}
\alias{R1mcMenu}
\alias{R12mcMenu}
\alias{randomnessFTest}
\alias{randomnessNTest}
\alias{S1mcMenu}
\alias{S12mcMenu}
\alias{SampleSizeZMenu}
\alias{sigma.test}
\alias{sigmaTest}
\alias{twolevelfactor.runs.test}
\alias{uchart1Menu}
\alias{uchart12Menu}
\alias{uchart2Menu}
\alias{xbar1mcMenu}
\alias{xbar12mcMenu}
\alias{xbar2mcMenu}
\alias{xbarone1Menu}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  RcmdrPlugin.UCA packages. 
  For user information ask for help about \code{\link[RcmdrPlugin.UCA]{RcmdrPlugin.UCA}}.}
\keyword{internal}
