\name{RcmdrPlugin.UCA-internal}
\title{RcmdrPlugin.UCA:  Some customization to Rcmdr made by R-UCA project}
\alias{Rcmdr.droplevels}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  RcmdrPlugin.UCA packages. 
  For user information ask for help about \code{\link[RcmdrPlugin.UCA]{RcmdrPlugin.UCA}}.}
\keyword{internal}
