\name{RcmdrPlugin.UCA-package}
\title{RcmdrPlugin.UCA: Some customization to Rcmdr made by R-UCA project}
\alias{RcmdrPlugin.UCA-package}
\alias{RcmdrPlugin.UCA}
\docType{package}
\encoding{UTF-8}
\description{
  This package provides customized function to Rcmdr (R Commander).

  All the customizations provides by this package will be propouse to be
  part of the Rcmdr package. All the customization included in the Rcmdr
  package will be droped from here.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.UCA\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2012-03-12\cr
License: \tab GPL version 3 or newer\cr
}
}

\author{
Manuel Munoz-Marquez <manuel.munoz@uca.es>
}

\keyword{package}
\seealso{
For more information see \code{\link[Rcmdr]{Rcmdr-package}}. 

Para ayuda en \enc{español}{espanol}, \enc{véase}{vease}
\link{RcmdrPlugin.UCA.es}. (For Spanish help see \link{RcmdrPlugin.UCA.es}.) 
}

\references{
[1] \url{http://knuth.uca.es/RcmdrPlugin.UCA}
}

