\name{RcmdrPlugin.UCA-package}
\title{RcmdrPlugin.UCA: Some customization to Rcmdr made by R-UCA project}
\alias{RcmdrPlugin.UCA-package}
\alias{RcmdrPlugin.UCA}
\docType{package}
\encoding{UTF-8}
\description{
  Some extension to Rcmdr (R Commander), randomness test, variance test
  for one normal sample and predictions using active model, to teach statistics in a first
  university course made by R-UCA project and used at University of
  Cadiz (UCA).

  All the customizations provides by this package will be propose to be
  part of the Rcmdr package. 
  All the customization included in the Rcmdr package will be dropped from here.  
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.UCA\cr
Type: \tab Package\cr
Version: \tab 4.3-4\cr
Date: \tab 2021-03-01\cr
License: \tab GPL version 3 or newer\cr
}
}

\examples{
# Examples of use are provided in the help of each option of menu.
}

\author{
Manuel Munoz-Marquez <manuel.munoz@uca.es>
}

\keyword{package}
\seealso{
For more information see \code{\link[Rcmdr]{Rcmdr-package}}. 

Para ayuda en \enc{español}{espanol}, \enc{véase}{vease}
\code{\link{RcmdrPlugin.UCA.es}}. (For Spanish help see \code{\link{RcmdrPlugin.UCA.es}}.) 
}

\references{
[1] \url{http://knuth.uca.es/RcmdrPlugin.UCA/}
}

