\name{RcmdrPlugin.UCA.es}
\title{RcmdrPlugin.UCA.es: Algunas adaptaciones para Rcmdr hechas por el
  proyecto R-UCA}
\alias{RcmdrPlugin.UCA.es}
\docType{package}
\encoding{UTF-8}

\description{
  Algunas extensiones a Rcmdr (R Commander), test de aleatoriedad, test
  de varianza para una muestra normal y
  predicciones utilizando el modelo activo, para la \enc{enseñanza}{ensenanza} de
  un primer curso universitario en \enc{estadística}{estadistica} hechas por el
  proyecto R-UCA y usadas en la Universidad de \enc{Cádiz}{Cadiz} (UCA).

  Todas las adaptaciones \enc{serán}{seran} propuestas para su \enc{incorporación}{incorporacion} a
  Rcmdr. Las adaptaciones que se incorporen al paquete Rcmdr se \enc{suprimirán}{suprimiran}
  de este paquete.
}
\details{
\tabular{ll}{
Paquete: \tab RcmdrPlugin.UCA\cr
Tipo: \tab Paquete\cr
Version: \tab 4.4-1\cr
Fecha: \tab 2021-04-02\cr
Licencia: \tab GPL (>= 3)\cr
}
}

\examples{
# Se proporcionan ejemplos de uso en la ayuda de cada opcion de menú.
}

\author{
Manuel \enc{Muñoz-Márquez}{Munoz-Marquez} <manuel.munoz@uca.es>
}

\keyword{package}
\seealso{
Para \enc{más}{mas} \enc{información}{informacion} \enc{véase}{vease} \code{\link[Rcmdr]{Rcmdr-package}}.
}

\references{
[1] \url{http://knuth.uca.es/RcmdrPlugin.UCA/}
}

