\name{RcmdrPlugin.UCA.es}
\title{RcmdrPlugin.UCA.es: Algunas adaptaciones para Rcmdr hechas por el
  proyecto R-UCA}
\alias{RcmdrPlugin.UCA.es}
\docType{package}
\encoding{UTF-8}

\description{
  Este paquete proporciona adaptaciones para Rcmdr (R Commander).

  Todas las adaptaciones \enc{serán}{seran} propuestas para su \enc{incorporación}{incorporacion} a
  Rcmdr. Las adaptaciones que se incorporen al paquete Rcmdr se \enc{suprimirán}{suprimiran}
  de este paquete.
}
\details{
\tabular{ll}{
Paquete: \tab RcmdrPlugin.UCA\cr
Tipo: \tab Paquete\cr
Version: \tab 1.0-1\cr
Fecha: \tab 2012-03-12\cr
Licencia: \tab GPL (>= 2)\cr
}
}

\author{
Manuel \enc{Muñoz-Márquez}{Munoz-Marquez} <manuel.munoz@uca.es>
}

\keyword{package}
\seealso{
Para \enc{más}{mas} \enc{información}{informacion} \enc{véase}{vease} \code{\link[Rcmdr]{Rcmdr-package}}.
}


\references{
[1] \url{http://knuth.uca.es/RcmdrPlugin.UCA}
}

